README: use correct capitalisation for "badge"
According to the Oxford manual of Style (ref. [1], p. 4), "do NOT use a
capital letter unless it is absolutely required".

The use of Badge is a absolutely obvious violation of this rule. Fix
this major issue by using a 'b' instead of a 'B' (ASCII offset 20
(0x14)).

This fix is in alignment with the Chicago manual of Style [2] Part II,
Style and Usage, section 8: "Names, Terms, and Titles of Work",
subsection 8.157 "Capitalization of titles of works–general principles".

[1] https://www.ox.ac.uk/sites/files/oxford/media_wysiwyg/University%20of%20Oxford%20Style%20Guide.pdf
[2] https://www.chicagomanualofstyle.org/book/ed17/part2/ch08/toc.html

Signed-off-by: Ralf Ramsauer <ralf@vmexit.de>